增加交易接口价格类型选项，修复重复移除持仓问题

+++ northstar-core/src/main/java/tech/xuanwu/northstar/core/domain/Account.java
@@ -118,7 +118,7 @@ public class Account implements IAccount{
 				//当持仓不变时，不更新数据库
 				return;
 			}
-			if(position.getPosition() == 0) {
+			if(positionMap.containsKey(position.getPositionId()) && position.getPosition() == 0) {
 				positionMap.remove(position.getPositionId());
 				positionRepo.removeById(position);
 				return;
